Refactor to a base class for query parameter transforms
